Move Applicant Portal to parallel route group
